+++ package.json
@@ -24,7 +24,7 @@
     "test:gui": "NODE_OPTIONS=--experimental-vm-modules npx jest --testMatch '**/gui/**/?(*.)+(spec|test).[jt]s?(x)' --coveragePathIgnorePatterns '/node_modules/' 'onnx/onnx_pb.js' --reporters default ./tests/retry-test.js ./tests/slow-test.js ./tests/gui-coverage-reporter.js --testTimeout=20000 --maxWorkers=100% --workerIdleMemoryLimit=100MB",
     "types": "npm run create-entry && find lib -name \\*.d.ts -delete && tsc ./lib/**.js ./lib/**/*.js --declaration --allowJs --emitDeclarationOnly",
     "document": "npm run types && typedoc --skipErrorChecking ./lib/**.d.ts",
-    "bundle": "node ./bundle_index.js",
+    "bundle": "node ./scripts/bundle_index.js",
     "minify": "npx esbuild `find . -name \\*.js | grep -E '^./(lib|js)'` ./css/*.css --minify --allow-overwrite --keep-names --outdir=. && html-minifier --collapse-whitespace --minify-css true --minify-js true -o ./index.html ./index.html",
     "build:gh-pages": "node ./scripts/inject-version.js && npm run document && find lib -name \\*.d.ts -delete && sed -i -e /^doc/d .gitignore && npm run bundle && npm run minify",
     "build:npm": "node ./scripts/inject-version.js && npm run types && npm run format"
@@ -71,7 +71,7 @@
     "jest": "^30.3.0",
     "onnxruntime-web": "^1.24.3",
     "playwright": "^1.58.2",
-    "rollup": "^4.59.0",
+    "rolldown": "^1.2.2",
     "typedoc": "^0.28.19",
     "typescript": "^6.0.3",
     "v8-to-istanbul": "^9.3.0",
